third_party/WebKit/LayoutTests/http/tests/inspector/network/network-xhr-post-payload.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>
function test()
{
    // Since this test could be run together with other inspector backend cache
    // tests, we need to reset size limits to default ones.
    InspectorTest.resetInspectorResourcesData(step1);
    var payload = "\u201AFoo\u201B";

    function step1()
    {
        InspectorTest.recordNetwork();
        InspectorTest.makeSimpleXHRWithPayload("POST", "resources/resource.php?foo", true, payload, step2);
    }

    function step2()
    {
        var request = InspectorTest.networkRequests().pop();
        InspectorTest.addResult(request.url);
        InspectorTest.assertEquals("foo", request.queryString(), "Unexpected resource query.");
        InspectorTest.assertEquals(payload, request.requestFormData, "Payload corrupted.");
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
<p>Tests XHR network resource payload is not corrupted by transcoding.</p>
</body>
</html>
