third_party/WebKit/LayoutTests/http/tests/inspector/network/network-xhr-post-payload-expected.txt

CONSOLE MESSAGE: line 6: XHR loaded: 1
Tests XHR network resource payload is not corrupted by transcoding.

http://127.0.0.1:8000/inspector/network/resources/resource.php?foo
